perf dso: make dsos list/rbtree namespace-aware
There is a bug with perf when looking for ELF files inside contaienrs.
If two or more contaienrs are running a process with the same path, but
with different binaries in that path (different versions of the same
program, for example), perf will only save the DSO for one of those
processes, and will use that DSO to translate symbols from both
processes, leading to incorrect symbol translation.

To prevent that, make the DSOs list/rbtree aware of namespace mount
information when looking up existing DSOs or adding new DSOs.